Networking improvements

+++ common/src/character/index.rs
@@ -1,7 +1,7 @@
 use serde::{Deserialize, Serialize};
 
 #[derive(Default, Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
-pub struct CharacterIndex(pub usize);
+pub struct CharacterIndex(pub u8);
 
 impl CharacterIndex {
     pub fn increment(&mut self) {
